+++ tests/fkooman/OAuth/Server/OAuthHelper.php
@@ -18,8 +18,9 @@
 namespace fkooman\OAuth\Server;
 
 use fkooman\Ini\IniReader;
+use PHPUnit_Framework_TestCase;
 
-class OAuthHelper extends \PHPUnit_Framework_TestCase
+class OAuthHelper extends PHPUnit_Framework_TestCase
 {
     protected $_tmpDb;
 
@@ -60,6 +61,7 @@ class OAuthHelper extends \PHPUnit_Framework_TestCase
                   "secret" => null,
                   "icon" => null,
                   "allowed_scope" => "read",
+                  "user_consent" => true,
                   "contact_email" => "foo@example.org",
                   "redirect_uri" => "http://localhost/php-oauth/unit/test.html",
                   "type" => "user_agent_based_application", );
@@ -70,6 +72,7 @@ class OAuthHelper extends \PHPUnit_Framework_TestCase
                   "secret" => "abcdef",
                   "icon" => null,
                   "allowed_scope" => "read write foo bar foobar",
+                  "user_consent" => true,
                   "contact_email" => null,
                   "redirect_uri" => "http://localhost/php-oauth/unit/test.html",
                   "type" => "web_application", );
@@ -80,6 +83,7 @@ class OAuthHelper extends \PHPUnit_Framework_TestCase
                   "icon" => null,
                   "allowed_scope" => "read",
                   "contact_email" => null,
+                  "user_consent" => true,
                   "redirect_uri" => "oauth://callback",
                   "type" => "native_application", );